+++ src/main/java/com/furniture/service/CouponService.java
@@ -4,11 +4,12 @@ import com.furniture.modal.Cart;
 import com.furniture.modal.Coupon;
 import com.furniture.modal.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CouponService {
 
-    Cart applyCoupon(String code, double orderValue, User user) throws Exception;
+    Cart applyCoupon(String code, BigDecimal orderValue, User user) throws Exception;
     Cart removeCoupon(String code, User user) throws Exception;
     Coupon findCouponById(Long id) throws Exception;
     Coupon createCoupon(Coupon coupon);
